\name{cdbkDf-class}
\Rdversion{1.1}
\docType{class}
\alias{cdbkDf-class}

\title{Class for data-level summaries}
\description{This is a virtual class for data-level summaries.}
\section{Extends}{
Class \code{"\linkS4class{cdbkObject}"}, directly.
}
\seealso{
This class is extended by the following classes:
\itemize{
  \item \code{\linkS4class{dfTech}}
}
}
\keyword{classes}
