\name{cdbkObject-class}
\docType{class}
\alias{cdbkObject-class}

\title{Package super-class}
\description{This is a virtual class that is a parent of all the objects
in the package \pkg{cdbk}.}
\section{Objects from the Class}{A virtual Class: No objects may be created
from it.}
\seealso{
Use \code{showClass("cdbkObject")} to list the sub-classes
defined in this package.
}
\examples{
showClass("cdbkObject")
}
