\name{cdbkVar-class}
\docType{class}
\alias{cdbkVar-class}

\title{Class for variable description objects}
\description{Virtual class for objects containing variable-level
documentation.}
\section{Extends}{
Class \code{"\linkS4class{cdbkObject}"}, directly.
}
\section{Methods}{
No methods defined with class "cdbkVar" in the signature.
}
\seealso{
See examples of the known sub-classes, e.g. \linkS4class{cdbkTech}.
}
\examples{
# this will also show defined sub-classes
showClass("cdbkVar")
}
\keyword{classes}
