\name{cdbkVector-class}
\docType{class}
\alias{cdbkVector-class}

\title{Class for vector-like objects}

\description{This is a virtual class for vector objects that are supported
as data for the variables. Currently these include the following classes:
"numeric", "character", "integer", "logical", "factor".}

\examples{
showClass("cdbkVector")
}
